Improved the design of WithAuthority for B2C by: - moving the extension class/method to a separate file - Adding an extension method both for the App creation and token acquisition.
Not related, also proposed a different approach for GetAccountsForPolicy. This which should not be part of the current API Review, but from another one to come (like next week :).
I hesitated into bringing it here, but this appeared so clearly that I thought we'd want to discuss it.

This way the sample for B2C is almost exactly the same as for AAD, but we just inject policies where needed.